+++ src/main/java/ru/itgirl/library_project/service/UsersServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.itgirl.library_project.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.itgirl.library_project.dto.UsersCreateDto;
 import ru.itgirl.library_project.dto.UsersDto;
@@ -9,18 +10,28 @@ import ru.itgirl.library_project.model.Users;
 import ru.itgirl.library_project.repository.UsersRepository;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class UsersServiceImpl implements UsersService {
     private final UsersRepository usersRepository;
 
     @Override
     public UsersDto getUsersById(Long id) {
-        Users users = usersRepository.findById(id).orElseThrow();
-        UsersDto usersDto = convertEntityToDto(users);
-        return usersDto;
+        log.info("Try to find user by id {}", id);
+        Optional<Users> users = usersRepository.findById(id);
+        if (users.isPresent()) {
+            UsersDto usersDto = convertEntityToDto(users.get());
+            log.info("User: {}", usersDto.toString());
+            return usersDto;
+        } else {
+            log.error("User with {id} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
@@ -43,7 +54,16 @@ public class UsersServiceImpl implements UsersService {
 
     @Override
     public void deleteUsers(Long id) {
-        usersRepository.deleteById(id);
+        log.info("Try to delete user with id {}", id);
+        Optional<Users> users = usersRepository.findById(id);
+        if (users.isPresent()) {
+            log.info("Found user with id {}", id);
+            usersRepository.deleteById(id);
+            log.info("Deleted user with id {}", id);
+        } else {
+            log.error("Failed to find user with id {}", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
